Move findIntersection out of Renderer into static function in PathIntersection class
